fix HSP64 feature
objinfo関数の修正
lldimを追加
#varint64を追加

+++ src/hspcmp/label.cpp
@@ -668,7 +668,7 @@ int CLabel::DumpHSPLabelById(int id, char* str, int option)
 	}
 
 	*p = 0;
-	return (p-str);
+	return (int)(p-str);
 }
